+++ examples/data-sources/sdwan_custom_control_topology_policy_definition/data-source.tf
@@ -1,0 +1,3 @@
+data "sdwan_custom_control_topology_policy_definition" "example" {
+  id = "f6b2c44c-693c-4763-b010-895aa3d236bd"
+}