+++ examples/resources/sdwan_custom_control_topology_policy_definition/import.sh
@@ -1,0 +1,1 @@
+terraform import sdwan_custom_control_topology_policy_definition.example "f6b2c44c-693c-4763-b010-895aa3d236bd"