+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `sdwan_cisco_secure_internet_gateway_feature_template` resource and data source
 - Add `sdwan_hub_and_spoke_topology_policy_definition` resource and data source
 - Add `sdwan_mesh_topology_policy_definition` resource and data source
+- Add `sdwan_custom_control_topology_policy_definition` resource and data source
 
 ## 0.2.0